pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "Бекенд для сайта визитки"
authors = [
    {name = "Alexey",email = "Starobogatov.a@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.11,<0.116.0)",
    "sqlalchemy[async] (>=2.0.39,<3.0.0)",
    "redis[async] (>=5.2.1,<6.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic[mypy] (>=2.10.6,<3.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "user-agents (>=2.2.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "mitmproxy (>=11.1.3,<12.0.0)",
    "locust (>=2.34.1,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "hypercorn (>=0.17.3,<0.18.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = "^(tests/|other_folder/).*"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
